[FIX] #58 - 그래프화면에도 색상 인덱스 적용

+++ app/src/main/java/com/woowahantechcamp/account_book/data/entity/StatisticEntity.kt
@@ -3,6 +3,6 @@ package com.woowahantechcamp.account_book.data.entity
 data class StatisticEntity(
     val categoryId: Int,
     val categoryTitle: String,
-    val color: Long,
+    val color: Int,
     val sum: Long
 )
